Fix DQM/PhysicsHWW by just ignoring jets with invalid parameters.
Fixes (silences) exceptions in relvals for 76x. See diff.

This module needs to be rewritten (or removed).
The code currently has no maintainter and it's in a pretty bad state.

In this particular case both hww.pfjets_p4 and hww.pfjets_JEC are filled
separately (and can have different sizes), even though
array index represents the same jet. This is definetly an error.